add the UITableView datasource and delegate types, and autobox them.  Also add a version of helloios that displays the demos in a table.
the coffeekit.coffee change is important.  the runtime doesn't marshal JSObjectRefs from objc objects if they aren't
specified in the parameter list of the function the runtime will be invoking.  the original protocol proxy code
used an args... splat to proxy all arguments transparent.  trouble is, this results in a function.length of 0, so none
of the parameters were being marshaled.  To fix it we switch on the function length and create specialized tramps for
a given argument list length.